+++ init/examples/base/islands/Keyboard.tsx
@@ -1,8 +1,0 @@
-import { Button } from 'lunchbox';
-// import { useRef } from "preact/hooks";
-// import { handleKeyboard } from "teclas";
-
-export default function Keyboard() {
-  // const ref = useRef<HTMLButtonElement>(null);
-  return <Button.Brand>Keyboard</Button.Brand>;
-}